Create pull_request_template.md (#74)
This PR closes #72. 

I've created a template for a meaningful PR so that we can use this for future pull requests. This is mandatory in the Senior Design Portfolio II. No code files were changed.

Please take a look at the template and let me know If you feel I should modify any sections of the template